# SwaggerDoc/Api.yaml
openapi: 3.0.0
info:
  description: Personal Goals
  version: "1"
  title: Personal Goals
  contact:
    email: lukas.slizik@gmail.com

paths:
  /activities:
    get:
      operationId: GetActivities
      responses:
        200:
          description: A JSON array of Activities
          content:
            application/json:
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Activity'
    post:
      operationId: CreateActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Activity'
      responses:
        201:
          description: New Activity Created

  /activities/{activityId}:
    get:
      operationId: GetActivityById
      parameters:
      - in: path
        name: activityId
        required: true
        schema:
          type: string
          format: guid
      responses:
        200:
          description: A JSON representation of an Activity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
    put:
      operationId: UpdateActivity
      parameters:
      - in: path
        name: activityId
        required: true
        schema:
          type: string
          format: guid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Activity'
      responses:
        204: 
          description: 'no content'
    delete:
      operationId: DeleteActivity
      parameters:
      - in: path
        name: activityId
        required: true
        schema:
          type: string
          format: guid
      responses:
        204:
          description: Activity Deleted

  /activities/{activityId}/goals:
    get:
      operationId: GetGoalsForActivity
      parameters:
      - in: path
        name: activityId
        required: true
        schema:
          type: string
          format: guid
      responses:
        200:
          description: A JSON array of Goals belonging to the Activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Goal'
    post:
      operationId: CreateGoal
      parameters:
      - in: path
        name: activityId
        required: true
        schema:
          type: string
          format: guid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Goal'
      responses:
        201:
          description: New Goal created

  /activities/{activityId}/goals/{goalId}:
    get:
      operationId: GetGoalById
      parameters:
      - in: path
        name: activityId
        required: true
        schema:
          type: string
          format: guid
      - in: path
        name: goalId
        required: true
        schema:
          type: string
          format: guid
      responses:
        200:
          description: A JSON representation of a Goal belonging to the Activity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
    put:
      operationId: UpdateGoal
      parameters:
      - in: path
        name: activityId
        required: true
        schema:
          type: string
          format: guid
      - in: path
        name: goalId
        required: true
        schema:
          type: string
          format: guid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Goal'
      responses:
        204:
          description: 'no content'

    delete:
      operationId: DeleteGoalById
      parameters:
      - in: path
        name: activityId
        required: true
        schema:
          type: string
          format: guid
      - in: path
        name: goalId
        required: true
        schema:
          type: string
          format: guid
      responses:
        204:
          description: Goal Deleted
      
  /activities/{activityId}/goals/{goalId}/achievements:
    get:
      operationId: GetAllAchievementsForGoal
      parameters:
        - in: path
          name: activityId
          required: true
          schema:
            type: string
            format: guid
        - in: path
          name: goalId
          required: true
          schema:
            type: string
            format: guid
      responses:
        200:
          description: A JSON Array  of all Achievements belonging to a Goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Achievement'
    post:
      operationId: CreateAchievement
      parameters:
      - in: path
        name: activityId
        required: true
        schema:
          type: string
          format: guid
      - in: path
        name: goalId
        required: true
        schema:
          type: string
          format: guid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Achievement'
      responses:
        201:
          description: Achievement created
      

  /activities/{activityId}/goals/{goalId}/achievments/{achievmentId}:
    get:
      operationId: GetAchievementById
      parameters:
        - in: path
          name: activityId
          required: true
          schema:
            type: string
            format: guid
        - in: path
          name: goalId
          required: true
          schema:
            type: string
            format: guid
        - in: path
          name: achievmentId
          required: true
          schema:
            type: string
            format: guid
      responses:
        200:
          description: A JSON representation of an Achievement belonging to a Goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Achievement'
    put:
      operationId: UpdateAchievement
      parameters:
        - in: path
          name: activityId
          required: true
          schema:
            type: string
            format: guid
        - in: path
          name: goalId
          required: true
          schema:
            type: string
            format: guid
        - in: path
          name: achievmentId
          required: true
          schema:
            type: string
            format: guid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Achievement'
      responses:
        200:
          description: A JSON representation of an Achievement belonging to a Goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Achievement'
    delete:
      operationId: DeleteAchievementById
      parameters:
      - in: path
        name: activityId
        required: true
        schema:
          type: string
          format: guid
      - in: path
        name: goalId
        required: true
        schema:
          type: string
          format: guid
      - in: path
        name: achievmentId
        required: true
        schema:
          type: string
          format: guid
      responses:
        204:
          description: Achievement was deleted

components:
  schemas:
    Goal:
      type: object
      properties:
        id:
          type: string
          format: guid
        target:
          type: number
        step:
          type: number
        unit:
          type: string
        multiplicator:
          type: number
          format: float
    Activity:
      type: object
      properties:
        id:
          type: string
          format: guid
        name:
          type: string
        description:
          type: string
    Achievement:
      type: object
      properties:
        id:
          type: string
          format: guid
        points:
          type: number
          format: float
        timePeriod:
          $ref: '#/components/schemas/TimePeriod'
    TimePeriod:
      type: object
      properties:
        id:
          type: string
          format: guid
        from:
          type: string
          format: date
        to:
          type: string
          format: date
# Added by API Auto Mocking Plugin
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/LukasSlizik/PersonalGoals/1